+++ packages/components/src/ActionList/ActionListItem.tsx
@@ -34,13 +34,29 @@ import { ActionListContext } from './ActionListContext'
 
 export interface ActionListItemProps
   extends CompatibleHTMLProps<HTMLDivElement> {
+  /**
+   *  The available actions for this item
+   */
   actions?: ReactNode
+  /**
+   *  Sets the tooltip text and a hidden text label for the actions button (accessible to assistive technology)
+   *  If unprovided by the user, a default string will be used instead
+   *  @default 'Options'
+   */
+  actionsTooltip?: string
+  /**
+   *  The id of this item
+   */
   id: string
+  /**
+   * A boolean indicating whether this item is selectable or not (the item will appear greyed out if true)
+   */
   disabled?: boolean
 }
 
 const ActionListItemInternal: FC<ActionListItemProps> = ({
   actions,
+  actionsTooltip = 'Options',
   children,
   className,
   disabled,
@@ -82,8 +98,8 @@ const ActionListItemInternal: FC<ActionListItemProps> = ({
   const itemActions = actions && (
     <div onClick={handleMenuClick}>
       <Menu hoverDisclosureRef={actionListItemRef}>
-        <MenuDisclosure>
-          <IconButton icon="DotsVert" label="Actions" size="medium" />
+        <MenuDisclosure tooltip={actionsTooltip}>
+          <IconButton icon="DotsVert" label={actionsTooltip} size="medium" />
         </MenuDisclosure>
         <MenuList>{actions}</MenuList>
       </Menu>